inference_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
torchvision==0.18.0
ultralytics==8.2.0      # YOLOv9 via ultralytics
Pillow==10.3.0
numpy==1.26.4
httpx==0.27.0
python-multipart==0.0.9
